Add Ethernet settings panel
Depends-Change-Id: I049320d81fe510245c6d92f48eaf25ea7d7b908f
Depends-change-Id: Ie1d2100e6324052057c267997d050b60761f009a
Depends-change-Id: I1766488371186391929b8dc6605cebf215c0baec
Depends-Change-Id: Id7ef6fcfafd36f1c288fa14f587a0d42d1996247
Depends-Change-Id: I39b79940e0495f915f602a16d99e1848bd3f2015
Author: Matt Gumbel <matthew.k.gumbel@linux.intel.com>
Author: Anton Cherkashyn <antonx.t.cherkashyn@intel.com>
Signed-off by: Anton Cherkashyn <antonx.t.cherkashyn@intel.com>
Change-Id: I09b78c87281495fda9564d3dd753c851bd8d89e4
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>

Resolve Ethernet Toggle Switch work not properly

Listening the Interface state change broadcast to modify
the status of ethernet

Issue: AXIA-2167
Depend-Change-Id: Ied3791b78bcc3c1515aa8da3280ad380a9fd1371
Change-Id: I23a46cc3c1fc2f2386a587df8b3c682c4b799d07
Signed-off-by: juntingwang <Junting.Wang@windriver.com>

Ethernet: Reconnect if the switch is checked

Enable the Settings application to reconnect using the
EthernetService if the switch is checked. This patch
also addresses the situation when one changes DHCP to
Static and a "disconnecting" event is received.

For: AXIA-2228
Change-Id: Ia8c6869302629fa7f297873d794b186de054dfcc
Signed-off-by: Constantin Musca <constantin.musca@intel.com>