Fix settings crash if camera params can't be retrieved (2)
Derp, even if it works with the borked TF101 cam atm,
forgot the case that could happen where mCamera will reach the finally while being null ;)

Since we are already catching a RuntimeException, we might also as well go all the way with that logic, since the NPE is a RuntimeException anyway.
And we don't really care about what failed currently, basically if it's broken we just consider it can't do torch.

Change-Id: I8f15b87a0a2bea47222d18f44065199c4390a9af